refactor: add Telegram env in Dockerfile, move telegram notification to utils
