+++ docker/Dockerfile
@@ -38,6 +38,8 @@ ENV TARGET_DB_PASSWORD=""
 ARG DEBIAN_FRONTEND=noninteractive
 ENV VERSION="v1.2.5"
 ENV BACKUP_CRON_EXPRESSION=""
+ENV TG_TOKEN=""
+ENV TG_CHAT_ID=""
 ARG WORKDIR="/config"
 ARG BACKUPDIR="/backup"
 ARG BACKUP_TMP_DIR="/tmp/backup"
