+++ utils/utils.go
@@ -7,10 +7,14 @@
 package utils
 
 import (
+	"bytes"
+	"encoding/json"
 	"fmt"
 	"github.com/spf13/cobra"
 	"io"
 	"io/fs"
+	"io/ioutil"
+	"net/http"
 	"os"
 	"strconv"
 )
@@ -153,13 +157,6 @@ func CheckEnvVars(vars []string) error {
 
 	return nil
 }
-func Welcome() {
-	fmt.Println()
-	fmt.Println("**********************************")
-	fmt.Println("            MySQL Backup          ")
-	fmt.Println("     @Copyright © 2024 jkaninda    ")
-	fmt.Println("***********************************")
-}
 
 // MakeDir create directory
 func MakeDir(dirPath string) error {
@@ -188,4 +185,68 @@ func GetIntEnv(envName string) int {
 		Error("Error: %v", err)
 	}
 	return ret
+}
+func sendMessage(msg string) {
+
+	Info("Sending notification... ")
+	chatId := os.Getenv("TG_CHAT_ID")
+	body, _ := json.Marshal(map[string]string{
+		"chat_id": chatId,
+		"text":    msg,
+	})
+	url := fmt.Sprintf("%s/sendMessage", getTgUrl())
+	// Create an HTTP post request
+	request, err := http.NewRequest("POST", url, bytes.NewBuffer(body))
+	if err != nil {
+		panic(err)
+	}
+	request.Header.Add("Content-Type", "application/json")
+	client := &http.Client{}
+	response, err := client.Do(request)
+	if err != nil {
+		panic(err)
+	}
+	code := response.StatusCode
+	if code == 200 {
+		Info("Notification has been sent")
+	} else {
+		body, _ := ioutil.ReadAll(response.Body)
+		Error("Message not sent, error: %s", string(body))
+	}
+
+}
+func NotifySuccess(fileName string) {
+	var vars = []string{
+		"TG_TOKEN",
+		"TG_CHAT_ID",
+	}
+
+	//Telegram notification
+	err := CheckEnvVars(vars)
+	if err == nil {
+		message := "PostgreSQL Backup \n" +
+			"Database has been backed up \n" +
+			"Backup name is " + fileName
+		sendMessage(message)
+	}
+}
+func NotifyError(error string) {
+	var vars = []string{
+		"TG_TOKEN",
+		"TG_CHAT_ID",
+	}
+
+	//Telegram notification
+	err := CheckEnvVars(vars)
+	if err == nil {
+		message := "PostgreSQL Backup \n" +
+			"An error occurred during database backup \n" +
+			"Error: " + error
+		sendMessage(message)
+	}
+}
+
+func getTgUrl() string {
+	return fmt.Sprintf("https://api.telegram.org/bot%s", os.Getenv("TG_TOKEN"))
+
 }
